refact: get rid of testData file and move data inline into addNotes test
Keeping things simple!
