+++ server/tests/tests/testInOrder.test.js
@@ -1,3 +1,4 @@
+// file to run tests in a specific order to validate api and database health
 import "./health.test"
 import "./addNotes.test"
 import "./getNotes.test"
